Use composition for transaction browser components Use shadcn buttons Add stories
