+++ src/app/wallet/all/page.tsx
@@ -2,6 +2,6 @@ import React from 'react';
 
 import { TransactionsBrowser } from './transactions-browser';
 
-export default function TxnsBrowserPage() {
+export default function WalletAllPage() {
   return <TransactionsBrowser />;
 }
