--- src/app/wallet/all/transactions-table.tsx
@@ -1,220 +1,0 @@
-import { LinkExternal02 } from '@untitled-ui/icons-react';
-import { ArrowDown, ChevronDown, Plus, Search, Send } from 'lucide-react';
-import React, { useState } from 'react';
-import { DateRange } from 'react-day-picker';
-
-import {
-  formatCurrency,
-  formatTransactionDate,
-  formatWalletAddress,
-} from '@/app/wallet/utils';
-import useCopy from '@/components/common/copy-button/use-copy';
-import { DatePickerWithRange } from '@/components/ui/date-range-picker';
-import { Skeleton } from '@/components/ui/skeleton';
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipProvider,
-  TooltipTrigger,
-} from '@/components/ui/tooltip';
-import { env } from '@/lib/env';
-
-export interface TableTransaction {
-  account: string;
-  date: string;
-  amount: string;
-  hash: string;
-  type: 'Sent' | 'Received' | 'Self-transfer';
-}
-
-interface TransactionsTableProps {
-  transactions: TableTransaction[];
-  searchValue?: string;
-  dateRange?: DateRange | undefined;
-  isLoading?: boolean;
-  onSearchChange?: (value: string) => void;
-  onDateRangeChange?: (range: DateRange | undefined) => void;
-  onClearAll?: () => void;
-  minimalMode?: boolean;
-}
-
-const TransactionsTable: React.FC<TransactionsTableProps> = ({
-  transactions,
-  searchValue = '',
-  dateRange,
-  isLoading = false,
-  onSearchChange,
-  onDateRangeChange,
-  onClearAll,
-  minimalMode = false,
-}) => {
-  const handleTxnClick = (hash: string) => {
-    window.open(`${env.NEXT_PUBLIC_BASE_EXPLORER}/tx/${hash}`, '_blank');
-  };
-
-  return (
-    <>
-      {isLoading ? (
-        <Skeleton className='h-24 rounded-3xl bg-gray-300 shadow-md' />
-      ) : (
-        <div className='bg-card relative min-h-full rounded-2xl p-6 shadow-md'>
-          {!minimalMode && (
-            <div className='mb-4 grid grid-cols-12 gap-4'>
-              <div className='relative col-span-12 lg:col-span-8'>
-                <div className='pointer-events-none absolute inset-y-0 left-0 flex items-center pl-3'>
-                  <Search className='size-5' />
-                </div>
-                <input
-                  type='text'
-                  placeholder='Search'
-                  value={searchValue}
-                  onChange={(e) =>
-                    onSearchChange && onSearchChange(e.target.value)
-                  }
-                  className='w-full rounded-md border border-gray-300 px-10 py-2 focus:outline-none focus:ring-2 focus:ring-blue-500'
-                />
-              </div>
-
-              <div className='relative col-span-12 md:col-span-4 lg:col-span-2'>
-                {onDateRangeChange && (
-                  <DatePickerWithRange
-                    date={dateRange}
-                    onDateChange={onDateRangeChange}
-                    triggerButton={
-                      <button className='text-md text-muted-foreground flex w-full items-center justify-center rounded-md border border-gray-300 px-4 py-2 focus:outline-none'>
-                        Date
-                        <ChevronDown className='ml-1 size-4' />
-                      </button>
-                    }
-                  />
-                )}
-              </div>
-              <div className='relative col-span-12 md:col-span-4 lg:col-span-2'>
-                <button
-                  className='text-md text-muted-foreground flex w-full items-center justify-center rounded-md border border-gray-300 px-4 py-2 focus:outline-none'
-                  onClick={onClearAll}
-                >
-                  Clear All
-                </button>
-              </div>
-            </div>
-          )}
-          <table className='min-w-full divide-y divide-gray-200'>
-            <thead className='border-b'>
-              <tr>
-                <th
-                  scope='col'
-                  className='w-2/5 py-3 text-left text-xs font-medium text-gray-500'
-                >
-                  To/From
-                </th>
-                <th
-                  scope='col'
-                  className='w-1/5 py-3 text-left text-xs font-medium text-gray-500'
-                >
-                  Date
-                </th>
-                <th
-                  scope='col'
-                  className='w-1/5 py-3 text-right text-xs font-medium text-gray-500'
-                >
-                  Amount
-                </th>
-              </tr>
-            </thead>
-            <tbody className='bg-white'>
-              {transactions &&
-                transactions.map((transaction, index) => (
-                  <tr key={index}>
-                    <td className='w-2/5 whitespace-nowrap py-4'>
-                      <div className='flex items-center'>
-                        <div className='h-12 w-12 flex-shrink-0'>
-                          <span className='flex h-12 w-12 items-center justify-center rounded-full bg-[#D7D7D7]'>
-                            {transaction.type === 'Sent' && (
-                              <ArrowDown className='size-6 text-white' />
-                            )}
-                            {transaction.type === 'Received' && (
-                              <Send className='size-6 text-white' />
-                            )}
-                            {transaction.type === 'Self-transfer' && (
-                              <Plus className='size-6 text-white' />
-                            )}
-                          </span>
-                        </div>
-                        <div className='ml-4'>
-                          <AddressText address={transaction.account} />
-                          <div className='mt-1 text-xs text-[#595B5A]'>
-                            {transaction.type === 'Self-transfer'
-                              ? 'Added'
-                              : transaction.type}
-                          </div>
-                        </div>
-                      </div>
-                    </td>
-                    <td className='w-1/5 whitespace-nowrap py-4'>
-                      <div className='text-sm font-medium text-gray-900'>
-                        {formatTransactionDate(transaction.date)}
-                      </div>
-                    </td>
-                    <td
-                      className='w-1/5 cursor-pointer whitespace-nowrap py-4'
-                      onClick={() => handleTxnClick(transaction.hash)}
-                    >
-                      <div className='flex items-center justify-end gap-2'>
-                        <div className='text-sm font-medium'>
-                          {transaction.type === 'Sent' ? '-' : '+'}{' '}
-                          {formatCurrency(transaction.amount)} USDC
-                        </div>
-                        <LinkExternal02 className='size-[1.125rem] text-[#595B5A]' />
-                      </div>
-                    </td>
-                  </tr>
-                ))}
-            </tbody>
-          </table>
-        </div>
-      )}
-    </>
-  );
-};
-
-export default TransactionsTable;
-
-// Tooltip state inspired by https://github.com/shadcn-ui/ui/issues/86#issuecomment-2241817826
-// TODO: Revisit "copy" flash when tooltip is closing (likely due to fade out animation)
-/** Local component to display formatted address and allow copying */
-const AddressText = ({ address }: { address: string }) => {
-  const formattedAddress = formatWalletAddress(address);
-  const { isCopied, handleClick } = useCopy(address);
-  const [open, setOpen] = useState(false);
-
-  return (
-    <TooltipProvider>
-      <Tooltip open={open || isCopied}>
-        <TooltipTrigger asChild>
-          <button
-            type='button'
-            className='cursor-pointer'
-            onClick={() => {
-              setOpen(!open);
-              handleClick();
-            }}
-            onMouseEnter={() => setOpen(true)}
-            onMouseLeave={() => setOpen(false)}
-            onTouchStart={() => setOpen(!open)}
-            onKeyDown={(e) => {
-              e.preventDefault();
-              if (e.key === 'Enter') {
-                setOpen(!open);
-                handleClick();
-              }
-            }}
-          >
-            {formattedAddress}
-          </button>
-        </TooltipTrigger>
-        <TooltipContent>{isCopied ? 'Copied!' : 'Copy'}</TooltipContent>
-      </Tooltip>
-    </TooltipProvider>
-  );
-};
